requirements.txt
fastapi[standard]
pydantic
uvicorn
bson
pymongo
python-docx